Openearthtool/python/applications/wps/brl_modelling/requirements.txt
flask-cors==3.0
flask==1.0
geojson==2.5
geoserver-restconfig==1.0.2
lxml==4.2
markupsafe==1.0
numpy==1.17
owslib==0.18
pandas==0.23
#proj4==5.2
psycopg2-binary==2.8.4
pyproj==1.9.6
pyshp==2.0
pywps==4.2
rasterio==1.1
scipy==1.1
shapely==1.6
simplejson==3.16
singledispatch==3.4
sqlalchemy==1.3
uwsgi==2.0
configparser==4.0.2
cytoolz==0.10
imod==0.8